Array Manipulator
Write a function that receives an array of integers and an array of string commands and executes them over the
array. The commands are as follows:
 add <index> <element> – adds element at the specified index (elements right from this position
inclusively are shifted to the right).
 addMany <index><element 1> <element 2> … <element n> – adds a set of elements at
the specified index.
 contains <element> – prints the index of the first occurrence of the specified element (if exists) in
the array or -1 if the element is not found.
 remove <index> – removes the element at the specified index.
 shift <positions> – shifts every element of the array the number of positions to the left (with
rotation).
o
For example, [1, 2, 3, 4, 5] -> shift 2 -> [3, 4, 5, 1, 2]
 sumPairs – sums the elements in the array by pairs (first + second, third + fourth, …).
o
For example, [1, 2, 4, 5, 6, 7, 8] -> [3, 9, 13, 8].
 print – stop receiving more commands and print the last state of the array in the following format:
`[ {element1}, {element2}, …elementN} ]`
Note: The elements in the array must be joined by comma and space (, ).