String Game
Create a program that executes changes over a string. On the first line, you are going to receive the string. On the following lines, you will be receiving commands until the "Done" command. There are six possible commands:
"Change {char} {replacement}"
Replace all occurrences of the char with the given replacement, then print the string.
"Includes {substring}"
Check if the string includes the given substring with and print "True" or "False".
"End {substring}"
Check if the string ends with the given substring and print "True" or "False".
"Uppercase"
Make the whole string uppercased, then print it.
"FindIndex {char}"
Find the index of the first occurrence of the given char, then print it.
"Cut {startIndex} {count}"
Remove all characters from the string, except those starting from the given start index and the next count of characters. Print only the cut chars.
Input
On the first line, you are going to receive the string.
On the following lines, until the "Done" command is received, you will be receiving commands.
All commands are case-sensitive.
The input will always be valid.
Output
Print the output of every command in the format described above.