Boss Rush
Create a program that checks if inputs are valid. On the first line, you will receive a number that indicates how many inputs you will receive on the following lines.

You will read lines with a boss name and title, and you should check if they are valid, considering the following rules:
Boss - the name should be
In upper case letters
Minimum four letters long
Surrounded by "|"
Title – should:
It contains exactly 2 words, and they contain only alphabetical letters and 1 whitespace between them.
Surrounded by "#"
The name and title should be split by a single ":"
Example for a valid input:  |GEORGI|:#Lead architect#
If the input is valid. Print in the following format:

"{boss name}, The {title}
>> Strength: {length of the name}
>> Armor: {length of the title}"

If the input is invalid, print "Access denied!"
Input / Constraints
On the 1st line, you will receive the number of inputs.
On the following n lines, you will have to check if a boss has a valid name and title.
Output
Print the output with the format described above.